+++ src/app/src/util/util.facilitiesCSV.js
@@ -1,9 +1,10 @@
 /* eslint-disable camelcase */
 import isArray from 'lodash/isArray';
 import get from 'lodash/get';
+import moment from 'moment';
 
 import { joinDataIntoCSVString } from './util';
-import { PPE_FIELD_NAMES } from './constants';
+import { PPE_FIELD_NAMES, EXTENDED_FIELD_TYPES } from './constants';
 
 export const csvHeaders = Object.freeze([
     'oar_id',
@@ -14,6 +15,136 @@ export const csvHeaders = Object.freeze([
     'lat',
     'lng',
 ]);
+
+const formatAttribution = ({ value, updated_at, contributor_name }) =>
+    !contributor_name
+        ? value
+        : `${value} (${contributor_name} on ${moment(updated_at).format(
+              'LL',
+          )})`;
+
+const formatNumberOfWorkers = ({
+    value: { max, min },
+    updated_at,
+    contributor_name,
+}) =>
+    formatAttribution({
+        value: max === min ? `${max}` : `${min}-${max}`,
+        contributor_name,
+        updated_at,
+    });
+
+const formatParentCompany = ({ value, contributor_name, updated_at }) =>
+    formatAttribution({
+        value: EXTENDED_FIELD_TYPES.find(
+            f => f.fieldName === 'parent_company',
+        ).formatValue(value),
+        contributor_name,
+        updated_at,
+    });
+
+const formatFacilityType = (
+    [m, r],
+    {
+        value: { matched_values = [], raw_values = [] },
+        updated_at,
+        contributor_name,
+    },
+) => {
+    const matches = matched_values.map(values =>
+        formatAttribution({
+            value: values[2],
+            contributor_name,
+            updated_at,
+        }),
+    );
+    let raw =
+        typeof raw_values === 'string' ? raw_values.split('|') : raw_values;
+    raw = raw.map(value =>
+        formatAttribution({
+            value,
+            contributor_name,
+            updated_at,
+        }),
+    );
+    return [m.concat(matches), r.concat(raw)];
+};
+
+const formatProcessingType = (
+    [m, r],
+    {
+        value: { matched_values = [], raw_values = [] },
+        updated_at,
+        contributor_name,
+    },
+) => {
+    const matches = matched_values.map(values =>
+        formatAttribution({
+            value: values[3],
+            contributor_name,
+            updated_at,
+        }),
+    );
+    let raw =
+        typeof raw_values === 'string' ? raw_values.split('|') : raw_values;
+    raw = raw.map(value =>
+        formatAttribution({
+            value,
+            contributor_name,
+            updated_at,
+        }),
+    );
+    return [m.concat(matches), r.concat(raw)];
+};
+
+const formatProductType = ({ value, contributor_name, updated_at }) =>
+    value.raw_values
+        .map(v =>
+            formatAttribution({
+                value: v,
+                contributor_name,
+                updated_at,
+            }),
+        )
+        .join('|');
+
+const formatExtendedFields = fields => {
+    const numberOfWorkers = fields.number_of_workers
+        .map(formatNumberOfWorkers)
+        .join('|');
+    const parentCompany = fields.parent_company
+        .map(formatParentCompany)
+        .join('|');
+    const [
+        facilityTypeMatches,
+        facilityTypeRaw,
+    ] = fields.facility_type.reduce(formatFacilityType, [[], []]);
+    const [
+        processingTypeMatches,
+        processingTypeRaw,
+    ] = fields.processing_type.reduce(formatProcessingType, [[], []]);
+    const productType = fields.product_type.map(formatProductType).join('|');
+    const columns = [
+        numberOfWorkers,
+        parentCompany,
+        facilityTypeMatches.join('|'),
+        facilityTypeRaw.join('|'),
+        processingTypeMatches.join('|'),
+        processingTypeRaw.join('|'),
+        productType,
+    ];
+    return columns;
+};
+
+const extendedFieldHeaders = [
+    'number_of_workers',
+    'parent_company',
+    'facility_type',
+    'facility_type_raw',
+    'processing_type',
+    'processing_type_raw',
+    'product_type',
+];
 
 export const createFacilityRowFromFeature = (feature, options) => {
     const {
@@ -26,6 +157,7 @@ export const createFacilityRowFromFeature = (feature, options) => {
             contributors,
             is_closed,
             contributor_fields,
+            extended_fields,
         },
         geometry: {
             coordinates: [lng, lat],
@@ -50,25 +182,41 @@ export const createFacilityRowFromFeature = (feature, options) => {
         options && options.includeClosureFields
             ? [is_closed ? 'CLOSED' : null]
             : [];
-    const contributorFields =
-        options && options.isEmbedded
-            ? contributor_fields.map(field => field.value)
+    let contributorFields = [];
+    if (options && options.isEmbedded) {
+        contributorFields = options.includeExtendedFields
+            ? contributor_fields.filter(
+                  field => !extendedFieldHeaders.includes(field.fieldName),
+              )
+            : contributor_fields;
+        contributorFields = contributorFields.map(field => field.value);
+    }
+
+    const extendedFields =
+        options && options.includeExtendedFields
+            ? formatExtendedFields(extended_fields)
             : [];
 
     return Object.freeze(
         [oar_id, name, address, country_code, country_name, lat, lng]
             .concat(contributorData)
             .concat(ppeFields)
-            .concat(closureFields)
-            .concat(contributorFields),
+            .concat(contributorFields)
+            .concat(extendedFields)
+            .concat(closureFields),
     );
 };
 
 export const makeFacilityReducer = options => (acc, next) =>
     acc.concat([createFacilityRowFromFeature(next, options)]);
 
-const getContributorFieldHeaders = facilities => {
-    const fields = get(facilities, '[0].properties.contributor_fields', []);
+const getContributorFieldHeaders = (facilities, options) => {
+    let fields = get(facilities, '[0].properties.contributor_fields', []);
+    if (options && options.isEmbedded) {
+        fields = fields.filter(
+            field => !extendedFieldHeaders.includes(field.fieldName),
+        );
+    }
     return fields.map(field => field.label);
 };
 
@@ -80,11 +228,16 @@ export const makeHeaderRow = (options, facilities = []) => {
     if (options && options.includePPEFields) {
         headerRow = headerRow.concat(PPE_FIELD_NAMES);
     }
+    if (options && options.isEmbedded) {
+        headerRow = headerRow.concat(
+            getContributorFieldHeaders(facilities, options),
+        );
+    }
+    if (options && options.includeExtendedFields) {
+        headerRow = headerRow.concat(extendedFieldHeaders);
+    }
     if (options && options.includeClosureFields) {
         headerRow = headerRow.concat(['is_closed']);
-    }
-    if (options && options.isEmbedded) {
-        headerRow = headerRow.concat(getContributorFieldHeaders(facilities));
     }
     return [headerRow];
 };